v1.5.4: Restructure nav - add Errors and GIT submenus
Navigation restructure:
- Add Errors submenu with API and Trends pages
- Add GIT submenu with Pipeline and Repos pages
- Simplify Infrastructure to Deployments (Vercel only)
- Remove broken Maintenance tools (no backend APIs exist)

Changes:
- Extract API Errors tab to /dashboard/errors/api
- Extract Error Trends tab to /dashboard/errors/trends
- Move Git Pipeline to /dashboard/git/pipeline
- Extract Git Repos tab to /dashboard/git/repos
- Add collapsible submenus in AdminLayout
- Remove maintenance pages (non-functional)
- Add PROD deploy protection rules to CLAUDE.md
